benchmarks/framework-shootout/apps/django_app/requirements.txt
django==5.1.*
gunicorn==23.*
psycopg2-binary==2.9.*
orjson==3.*
